Moved download_site method to DownloadHandler.py
