+++ lib/DownloadHandler.py
@@ -11,13 +11,15 @@ from abc import ABC, abstractmethod
 from datetime import timedelta
 from io import BytesIO
 from itertools import islice
-from multiprocessing import Queue
 
 import requests
+from dateutil.parser import parse as parse_datetime
 from tqdm import tqdm
 from tqdm.contrib.concurrent import thread_map
 
+from lib.DatabaseLayer import getInfo
 from lib.LogHandler import LogHandler
+from lib.redis_q import RedisQueue
 
 thread_local = threading.local()
 logging.setLoggerClass(LogHandler)
@@ -38,11 +40,13 @@ class DownloadHandler(ABC):
 
         self.prefix = prefix
 
+        self.queue = RedisQueue(name=self.feed_type)
+
+        self.file_queue = RedisQueue(name="files")
+
         self.progress_bar = None
 
         self.last_modified = None
-
-        self.file_queue = Queue()
 
         self.logger = logging.getLogger("DownloadHandler")
 
@@ -58,14 +62,12 @@ class DownloadHandler(ABC):
             thread_local.session = requests.Session()
         return thread_local.session
 
-    def process_downloads(self, sites, q, collection):
+    def process_downloads(self, sites, collection):
         """
         Method to download and process files
 
         :param sites: List of file to download and process
         :type sites: list
-        :param q: Multiprocessing Queue reference
-        :type q: Queue
         :param collection: Mongodb Collection name
         :type collection: str
         :return:
@@ -78,7 +80,7 @@ class DownloadHandler(ABC):
 
         thread_map(self.download_site, sites, desc="Downloading and processing files")
 
-        self._process_queue_to_db(worker_size, q, collection=collection)
+        self._process_queue_to_db(worker_size, collection=collection)
 
         self.logger.info(
             "Duration: {}".format(timedelta(seconds=time.time() - start_time))
@@ -98,14 +100,14 @@ class DownloadHandler(ABC):
         for i in range(0, len(lst), number):
             yield lst[i: i + number]
 
-    def _handle_queue_progressbar(self, q, description):
+    def _handle_queue_progressbar(self, description):
         """
         Method for handling the progressbar during queue processing
 
-        :param q: Multiprocessing Queue reference
-        :type q: Queue
+        :param description: Description for tqdm progressbar
+        :type description: str
         """
-        max_len = q.qsize()
+        max_len = self.queue.qsize()
 
         pbar = tqdm(total=max_len, desc=description)
         not_Done = True
@@ -115,7 +117,7 @@ class DownloadHandler(ABC):
 
         while not_Done:
 
-            current_q_len = q.qsize()
+            current_q_len = self.queue.qsize()
 
             if x % 10 == 0:
                 # log stats the first cycle and every 10th cycle thereafter
@@ -147,28 +149,26 @@ class DownloadHandler(ABC):
 
         pbar.close()
 
-    def _process_queue_to_db(self, max_workers, q, collection):
+    def _process_queue_to_db(self, max_workers, collection):
         """
         Method to write the queued database transactions into the database given a Queue reference and Collection name
 
         :param max_workers: Max amount of worker processes to use; defaults to min(32, os.cpu_count() + 4)
         :type max_workers: int
-        :param q: Multiprocessing Queue reference
-        :type q: Queue
         :param collection: Mongodb Collection name
         :type collection: str
         """
 
-        pbar = mp.Process(target=self._handle_queue_progressbar, args=(q, "Processing queue"))
+        pbar = mp.Process(target=self._handle_queue_progressbar, args=("Processing queue", ))
 
         processes = [
-            mp.Process(target=self._db_bulk_writer, args=(q, collection,))
+            mp.Process(target=self._db_bulk_writer, args=(collection, ))
             for _ in range(max_workers)
         ]
         for proc in processes:
             proc.start()
             # Put triggers in the Queue to tell the workers to exit their for-loop
-            q.put(self._end)
+            self.queue.put(self._end)
 
         pbar.start()
 
@@ -177,12 +177,10 @@ class DownloadHandler(ABC):
 
         pbar.join()
 
-    def _db_bulk_writer(self, output_queue, collection, threshold=1000):
+    def _db_bulk_writer(self, collection, threshold=1000):
         """
         Method to act as worker for writing queued entries into the database
 
-        :param output_queue: Multiprocessing Queue reference
-        :type output_queue: Queue
         :param collection: Mongodb Collection name
         :type collection: str
         :param threshold: Batch size threshold; defaults to 1000
@@ -190,9 +188,8 @@ class DownloadHandler(ABC):
         """
         from lib.Config import Configuration as conf
         database = conf.getMongoConnection()
-        items = iter(output_queue.get, self._end)
 
-        for batch in iter(lambda: list(islice(items, threshold)), []):
+        for batch in iter(lambda: list(islice(self.queue, threshold)), []):
             database[collection].bulk_write(batch)
 
     def store_file(self, response_content, content_type, url):
@@ -253,12 +250,47 @@ class DownloadHandler(ABC):
 
         return wd, filename
 
-    @abstractmethod
-    def download_site(self, **kwargs):
-        raise NotImplementedError
+    def download_site(self, url):
+        self.logger.debug("Downloading from url: {}".format(url))
+        session = self.get_session()
+        with session.get(url) as response:
+            self.last_modified = parse_datetime(
+                response.headers["last-modified"], ignoretz=True
+            )
+
+            self.logger.debug(
+                "Last {} modified value: {} for URL: {}".format(self.feed_type, self.last_modified, url)
+            )
+
+            i = getInfo(self.feed_type.lower())
+
+            if i is not None:
+                if self.last_modified == i["last-modified"]:
+                    self.logger.info(
+                        "{}'s are not modified since the last update".format(
+                            self.feed_type
+                        )
+                    )
+                    sys.exit(0)
+
+            content_type = response.headers["content-type"]
+
+            self.logger.debug("URL: {} fetched Content-Type: {}".format(url, content_type))
+
+            wd, filename = self.store_file(
+                response_content=response.content, content_type=content_type, url=url
+            )
+
+            if filename is not None:
+                self.file_to_queue((wd, filename))
+            else:
+                self.logger.error(
+                    "Unable to retrieve a filename; something went wrong during the storing of the file"
+                )
+                sys.exit(1)
 
     @abstractmethod
-    def file_to_queue(self, **kwargs):
+    def file_to_queue(self, *args):
         raise NotImplementedError
 
     @abstractmethod
