+++ lib/XMLFileHandler.py
@@ -1,9 +1,5 @@
-import sys
 from abc import abstractmethod
 
-from dateutil.parser import parse as parse_datetime
-
-from lib.DatabaseLayer import getInfo
 from lib.DownloadHandler import DownloadHandler
 
 
@@ -14,45 +10,6 @@ class XMLFileHandler(DownloadHandler):
     def __repr__(self):
         """ return string representation of object """
         return "<< XMLFileHandler:{} >>".format(self.feed_type)
-
-    def download_site(self, url):
-        self.logger.debug("Downloading from url: {}".format(url))
-        session = self.get_session()
-        with session.get(url) as response:
-            self.last_modified = parse_datetime(
-                response.headers["last-modified"], ignoretz=True
-            )
-
-            self.logger.debug(
-                "Last {} modified value: {} for URL: {}".format(self.feed_type, self.last_modified, url)
-            )
-
-            i = getInfo(self.feed_type.lower())
-
-            if i is not None:
-                if self.last_modified == i["last-modified"]:
-                    self.logger.info(
-                        "{}'s are not modified since the last update".format(
-                            self.feed_type
-                        )
-                    )
-                    sys.exit(0)
-
-            content_type = response.headers["content-type"]
-
-            self.logger.debug("URL: {} fetched Content-Type: {}".format(url, content_type))
-
-            wd, filename = self.store_file(
-                response_content=response.content, content_type=content_type, url=url
-            )
-
-            if filename is not None:
-                self.file_to_queue((wd, filename))
-            else:
-                self.logger.error(
-                    "Unable to retrieve a filename; something went wrong during the storing of the file"
-                )
-                sys.exit(1)
 
     @abstractmethod
     def file_to_queue(self, *args):
